displayed the doctor notes in the patient profile
