+++ client/src/components/Create-patient-symptom.component.js
@@ -2,8 +2,8 @@ import React, { Component } from 'react';
 import axios from 'axios';
 import { ToastContainer, toast } from 'react-toastify';
 import 'react-toastify/dist/ReactToastify.css';
-
-import { Card, Button, CardTitle, CardText, Row, Col, UncontrolledCollapse, CardBody } from 'reactstrap';
+import DoctorNotes from './ProviderNoteDisplay.component'
+import { Card, Button, Row, Col } from 'reactstrap';
 
 
 const isFormValid = ({ formErrors, ...rest }) => {
@@ -37,7 +37,6 @@ export default class createSympotom extends Component {
       comment: '',
       updateNote:'none',
       immediateAttention: false,
-      doctorNoteArray: [],
       formErrors: { 
         valid : "Please enter a valid number",
         highTemp : "This is a high temperature. We will create an alert for your doctor",
@@ -54,16 +53,6 @@ export default class createSympotom extends Component {
       updateNote:'none',
       immediateAttention: false}
   }
-
-  async componentDidMount() {
-    //retrieves all the doctor notes
-    axios.get('http://localhost:5000/doctorNotes/' + this.props.patientId)
-        .then(response => {
-            console.log(response);
-            this.setState({doctorNoteArray : response.data});
-        });
-  };
-
   resetForm() {
     this.setState(this.baseState)
     // this.setState({
@@ -205,14 +194,6 @@ export default class createSympotom extends Component {
 
   render() {
     const { formErrors } = this.state;
-    const doctorNotes = this.state.doctorNoteArray.map(elem => {
-      return(
-        <div key={elem}>
-      <Card body key={elem}>
-          <CardText key={elem}>{elem}</CardText>
-       </Card>
-       </div>)
-  })
     return (
 
       <div>
@@ -302,23 +283,7 @@ export default class createSympotom extends Component {
             </Card>
           </Col>
           <Col sm="4">
-            <Card >
-              <CardTitle > <h3 className="text-center">Latest Doctor Notes</h3></CardTitle>
-              
-
-              
-                <Button className="viewAllDoctorNotesButton" color="primary" id="toggler" style={{ marginBottom: '1rem' }}>View All Doctor Notes</Button>
-                  <UncontrolledCollapse toggler="#toggler">
-                    <Card>
-                      <CardBody>
-                        {doctorNotes}
-                      </CardBody>
-                    </Card>
-                  </UncontrolledCollapse>
-
-                <p>Latest Note: {this.state.doctorNoteArray[this.state.doctorNoteArray.length - 1]}</p>
-
-            </Card>
+            <DoctorNotes patientId={this.props.patientId}/>
           </Col>
         </Row>
       </div>
