+++ client/src/components/PatientProfile.component.js
@@ -1,9 +1,9 @@
 import React, {Component} from 'react';
 import axios from 'axios';
 import CreateSymptoms from './CreateSymptoms.component';
-import { Button } from 'reactstrap';
+import { Button, Row, Col } from 'reactstrap';
 import {Link} from 'react-router-dom';
-
+import DoctorNotes from './ProviderNoteDisplay.component'
 
 export default class PatientProfile extends Component{ 
    constructor(props){
@@ -49,18 +49,13 @@ componentDidMount(){
 }
 
 render() {
-   const patientS=
-         <ul>{this.state.patientEntry.map((patientEntry) =>
-             <li key={patientEntry._id}>{patientEntry}
-             </li>
-             )}
-        </ul>
-
     const patientP =
             <ol>{this.state.patient}</ol>
 
-   
     return(
+        <div className="patientProfile">
+        <Row>
+            <Col sm="8">
         <div className = "container">
            <h1>Profile:</h1>
            <h4>Patient name: <i>{this.state.patientname}</i>,
@@ -69,7 +64,6 @@ render() {
            <h2>History:</h2>
                
           <CreateSymptoms patientId='5ecaabd07dfcc538bce811fc'/> 
-             <div>Last Patient Note from old schema: {patientS}</div>
              {this.props.isDoctor ? 
                 <div >
                     <Link to="/patient/doctorNotes" className="nav-link">
@@ -79,6 +73,14 @@ render() {
                 </div>
              : null }
 
+        </div>
+        </Col>
+        <Col sm="4">
+        <div>
+                <DoctorNotes patientId="5ecaabd07dfcc538bce811fc"/>
+        </div>
+        </Col>
+        </Row>
         </div>
     )
 }
